# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Game tracker app API",
        "title": "Game tracker API",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/platforms": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "platforms"
                ],
                "summary": "Get all platforms",
                "operationId": "get-all-platforms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Platforms"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "platforms"
                ],
                "summary": "Create a platform",
                "operationId": "create-platform",
                "parameters": [
                    {
                        "description": "Tag",
                        "name": "platform",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Platforms"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Platforms"
                        }
                    }
                }
            }
        },
        "/platforms/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "platforms"
                ],
                "summary": "Delete a platform",
                "operationId": "delete-platform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Platforms"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Platforms": {
            "type": "object",
            "properties": {
                "iconName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}